server/pyproject.toml
[tool.poetry]
name = "recipe-generator"
version = "0.1.0"
description = "Recipe generator using OpenAI API"
authors = ["Thibaud FRERE <thibaud.frere@gmail.com>"]
readme = "README.md"
packages = [
    { include = "recipe_generator" }
]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.12.0"
aiohttp = "^3.9.3"
aiofiles = "^23.2.1"
python-dotenv = "^1.0.1"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
beautifulsoup4 = "^4.12.2"
mechanicalsoup = "^1.3.0"
pydantic = "^2.5.2"
Pillow = "^10.2.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.4"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
pytest-aiohttp = "^1.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=recipe_generator --cov-report=term-missing" 